Corrections for 64 bits shifts.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@2818 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR0/HWACCMR0.cpp
@@ -417,7 +417,7 @@ HWACCMR0DECL(void) HWACCMR0DumpDescriptor(PX86DESCHC Desc, RTSEL Sel, const char
         u32Limit = u32Limit << PAGE_SHIFT | PAGE_OFFSET_MASK;
 
 #if HC_ARCH_BITS == 64
-    uint64_t    u32Base =  Desc->Gen.u32BaseHigh3 << 32ULL | Desc->Gen.u8BaseHigh2 << 24ULL | Desc->Gen.u8BaseHigh1 << 16ULL | Desc->Gen.u16BaseLow;
+    uint64_t    u32Base =  ((uintptr_t)Desc->Gen.u32BaseHigh3 << 32ULL) | Desc->Gen.u8BaseHigh2 << 24ULL | Desc->Gen.u8BaseHigh1 << 16ULL | Desc->Gen.u16BaseLow;
 
     Log(("%s %04x - %VX64 %VX64 - base=%VX64 limit=%08x dpl=%d %s\n", pszMsg,
          Sel, Desc->au64[0], Desc->au64[1], u32Base, u32Limit, Desc->Gen.u2Dpl, szMsg));
